+++ backend/routes/bot/utils/google_bucket.py
@@ -3,7 +3,7 @@ from datetime import timedelta
 from google.cloud import storage
 
 from backend.core.configs.config import config
-from backend.core.utils.gcp_auth import get_signing_credentials
+from backend.routes.google_bucket.utils import get_signing_credentials
 
 
 async def generate_download_url(storage_client: storage.Client, filename: str):
@@ -15,7 +15,7 @@ async def generate_download_url(storage_client: storage.Client, filename: str):
     """
     # Generate signed URL using impersonated credentials to avoid private key requirement
     impersonated_credentials = get_signing_credentials(config.VM_SERVICE_ACCOUNT_EMAIL)
-    
+
     blob = storage_client.bucket(config.BUCKET_NAME).blob(filename)
 
     signed_url = blob.generate_signed_url(
