+++ terraform/providers.tf
@@ -1,8 +1,8 @@
 # providers.tf
 
 provider "google" {
-    project = "nuspace-staging"
-    credentials = file("terraform.json")
-    region = "europe-central2"
-    zone = "europe-central2-a"
+    project = var.project_id
+    credentials = file(var.credentials_file)
+    region = var.region
+    zone = var.zone
 }
