--- terraform/terraform.tfvars
@@ -1,4 +1,0 @@
-push_endpoint = "https://stage.nuspace.kz/api/bucket/gcs-hook"
-push_auth_service_account_email = "nuspace-vm-sa@nuspace-staging.iam.gserviceaccount.com"
-push_auth_audience = "https://stage.nuspace.kz"
-subscription_suffix = "staging"
